requirements.txt
selene==2.0.0rc1
allure-pytest==2.13.1